U: Split models
